Migrate to multi-project build Move demos and playground to separate subproject
