+++ .gitignore
@@ -31,6 +31,7 @@ nb-configuration.xml
 # IntelliJ
 
 .idea/
+out/
 # if you remove the above rule, at least ignore user-specific stuff:
 # .idea/workspace.xml
 # .idea/tasks.xml
