--- VWorkflows-Core/build.gradle
@@ -1,66 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'application'
-apply plugin: 'idea'
-apply plugin: 'eclipse'
-
-apply from: 'jdkdetect.gradle'
-apply from: 'version-info.gradle'
-//apply from: 'maven-deployment.gradle'
-apply from: 'maven-deployment-local.gradle'
-
-// all projects must be comatible with the java version specified here
-// overriding this property in subprojects is not allowed
-sourceCompatibility = '1.8'
-[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
-
-//apply from: 'http://gradle-plugins.mihosoft.eu/latest/vlicenseheader.gradle'
-//repairHeaders.licenseHeaderText = new File(projectDir,'../license-template.txt')
-
-repositories {
-    
-    mavenLocal()
-    
-    mavenCentral()
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = '2.1'
-}
-
-
-dependencies {
-
-    compile "com.thoughtworks.xstream:xstream:1.4.7"
-    
-    println("Gradle uses JDK "+project.jdk)
-
-    // compile project('..:VRLInstrumentation')    
-    // compile fileTree (dir: 'path-to-libdir', includes: ['*.jar'])
-    
-    testCompile 'junit:junit:4.11'
-}
-
-
-mainClassName =  "eu.mihosoft.vrl.workflow.playground.Playground"
-
-// -- coverage support (jacoco/coveralls) --
-
-apply plugin: 'jacoco'
-apply plugin: 'com.github.kt3k.coveralls'
-
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath 'org.kt3k.gradle.plugin:coveralls-gradle-plugin:2.0.1'
-    }
-}
-
-jacocoTestReport {
-    reports {
-        xml.enabled = true // coveralls plugin depends on xml format report
-        html.enabled = true
-    }
-}
