--- VWorkflows-Core/settings.gradle
@@ -1,4 +1,0 @@
-rootProject.name = 'vworkflows-core'
-
-//include 'jfxtras-labs'
-//project(':jfxtras-labs').projectDir = new File(settingsDir, '../../jfxtras-lab')
