+++ VWorkflows-Demo/src/main/resources/eu/mihosoft/vrl/workflow/demo/NodeControlFXML.fxml
