+++ VWorkflows-Demo/vworkflows-demo.gradle
@@ -1,0 +1,12 @@
+apply plugin: 'application'
+apply from: rootProject.file('gradle/javafx-plugin.gradle')
+
+mainClassName= 'eu.mihosoft.vrl.workflow.demo.Main'
+
+javafx {
+    mainClass = mainClassName
+}
+
+dependencies {
+    compile project(':vworkflows-fx') 
+}
