--- VWorkflows-FX/build.gradle
@@ -1,70 +1,0 @@
-//apply from: 'http://dl.bintray.com/content/shemnon/javafx-gradle/javafx.plugin'
-
-//apply from: 'http://gradle-plugins.mihosoft.eu/latest/vlicenseheader.gradle'
-//repairHeaders.licenseHeaderText = new File(projectDir,'../license-template.txt')
-apply plugin: 'idea'
-apply plugin: 'eclipse'
-apply plugin:'application'
-
-apply from: 'version-info.gradle'
-apply from: 'jdkdetect.gradle'
-
-
-
-//apply from: 'lib/gradle/javafx.plugin'
-
-//apply from: 'maven-deployment.gradle'
-apply from: 'maven-deployment-local.gradle'
-
-if (!hasProperty('mainClass')) {
-    ext.mainClass = 'eu.mihosoft.vrl.workflow.demo.Main'
-}
-
-//applicationDefaultJvmArgs = ["-Xss515m"]
-mainClassName = mainClass
-
-repositories {
-    mavenCentral()
-    maven {
-        url "https://oss.sonatype.org/content/repositories/snapshots"
-    }
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = '2.1'
-}
-
-// all projects must be comatible with the java version specified here
-// overriding this property in subprojects is not allowed
-sourceCompatibility = '1.8'
-[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
-
-dependencies {
-    
-    //    compile "org.jfxtras:jfxtras-labs-8.0:8.0-r1-SNAPSHOT"
-    compile "com.thoughtworks.xstream:xstream:1.4.4"
-    if (project.jdk == 7) {
-//        compile fileTree (dir: 'lib/jfxtras/jdk7/', includes: ['*.jar'])
-        compile "org.jfxtras:jfxtras-labs:2.2-r6-SNAPSHOT"
-    } else if (project.jdk == 8) {
-        compile "org.jfxtras:jfxtras-labs:8.0-r1"
-//        compile "org.jfxtras:jfxtras-window:8.0-r1-SNAPSHOT"
-        //compile fileTree (dir: 'lib/jfxtras/jdk8/', includes: ['*.jar'])
-    }
-    
-    compile group: 'org.apache.commons', name:'commons-math3', version: '3.3'
-    compile group: 'org.apache.commons', name:'commons-math3', version: '3.3', classifier: 'sources'
-    
-    compile project(':vworkflows-core')    
-    // compile fileTree (dir: 'path-to-libdir', includes: ['*.jar'])
-}
-
-
-
-// javafx {
-//     mainClass "eu.mihosoft.vrl.workflow.demo.Main"
-// }
-
-// jfxDeploy {
-//    packaging = "image"
-// }
