--- VWorkflows-FX/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Oct 20 10:55:29 CEST 2014
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.1-bin.zip
