--- VWorkflows-FX/settings.gradle
@@ -1,6 +1,0 @@
-rootProject.name = 'vworkflows-fx'
-
-// include 'jfxtras-labs'
-// project(':jfxtras-labs').projectDir = new File(settingsDir, '../../jfxtras-lab')
-include 'vworkflows-core'
-project(':vworkflows-core').projectDir = new File(settingsDir, '../VWorkflows-Core')
