--- VWorkflows-FX/src/main/java/eu/mihosoft/vrl/workflow/demo/Main.java
@@ -1,113 +1,0 @@
-/*
- * Main.java
- * 
- * Copyright 2012-2013 Michael Hoffer <info@michaelhoffer.de>. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification, are
- * permitted provided that the following conditions are met:
- *
- *    1. Redistributions of source code must retain the above copyright notice, this list of
- *       conditions and the following disclaimer.
- *
- *    2. Redistributions in binary form must reproduce the above copyright notice, this list
- *       of conditions and the following disclaimer in the documentation and/or other materials
- *       provided with the distribution.
- * 
- * Please cite the following publication(s):
- *
- * M. Hoffer, C.Poliwoda, G.Wittum. Visual Reflection Library -
- * A Framework for Declarative GUI Programming on the Java Platform.
- * Computing and Visualization in Science, 2011, in press.
- *
- * THIS SOFTWARE IS PROVIDED BY Michael Hoffer <info@michaelhoffer.de> "AS IS" AND ANY EXPRESS OR IMPLIED
- * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
- * FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL Michael Hoffer <info@michaelhoffer.de> OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
- * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * The views and conclusions contained in the software and documentation are those of the
- * authors and should not be interpreted as representing official policies, either expressed
- * or implied, of Michael Hoffer <info@michaelhoffer.de>.
- */ 
-
-package eu.mihosoft.vrl.workflow.demo;
-
-import java.io.IOException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
-import javafx.scene.Scene;
-import javafx.scene.layout.StackPane;
-import javafx.scene.paint.Color;
-import javafx.scene.shape.Rectangle;
-import javafx.stage.Stage;
-import jfxtras.labs.util.event.MouseControlUtil;
-
-/**
- *
- * @author Michael Hoffer  &lt;info@michaelhoffer.de&gt;
- */
-public class Main extends Application {
-
-    /**
-     * The main() method is ignored in correctly deployed JavaFX application.
-     * main() serves only as fallback in case the application can not be
-     * launched through deployment artifacts, e.g., in IDEs with limited FX
-     * support. NetBeans ignores main().
-     *
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        launch(args);
-    }
-    private MainWindowFXMLController controller;
-
-    @Override
-    public void start(Stage primaryStage) {
-
-//        connectionTest();
-
-        StackPane canvas = new StackPane();
-
-        Scene scene = new Scene(canvas, 1600, 800);
-
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("MainWindowFXML.fxml"));
-
-        try {
-            fxmlLoader.load();
-        } catch (IOException ex) {
-            Logger.getLogger(Main.class.getName()).
-                    log(Level.SEVERE, null, ex);
-        }
-
-        controller = fxmlLoader.getController();
-
-        canvas.getChildren().add((Node) fxmlLoader.getRoot());
-
-        primaryStage.setTitle("VWorkflows Demo!");
-        primaryStage.setScene(scene);
-        primaryStage.show();
-
-        Rectangle rect = new Rectangle();
-        rect.setStroke(new Color(1, 1, 1, 1));
-        rect.setFill(new Color(0, 0, 0, 0.5));
-        
-//        MouseControlUtil.
-//                addSelectionRectangleGesture(controller.getRootPane(), rect);
-
-//        Playground.search01();
-//        
-//        VRLShell shell = VShell.showInteractivScenicView(canvas);
-//        
-//        controller.registerShell(shell);
-
-    }
-
-   
-}
