--- VWorkflows-FX/version-info.gradle
@@ -1,11 +1,0 @@
-apply from: 'jdkdetect.gradle'
-apply plugin: 'maven'
-
-// maven deployment properties
-if (project.jdk == 8) {
-    group = 'eu.mihosoft.vrl.workflow-8.0'
-} else {
-    group = 'eu.mihosoft.vrl.workflow'
-}
-version = '0.1-r2-SNAPSHOT'
-
