+++ VWorkflows-FX/vworkflows-fx.gradle
@@ -1,0 +1,21 @@
+apply from: rootProject.file('gradle/javafx.gradle')
+
+// mainClassName = 'eu.mihosoft.vrl.workflow.demo.Main'
+
+repositories {
+    maven { url 'https://oss.sonatype.org/content/repositories/snapshots' }
+}
+
+dependencies {
+    compileOnly project.files(project.jfxrtLocation)
+
+    compile project(':vworkflows-core') 
+
+    if (project.jdk == 7) {
+        compile "org.jfxtras:jfxtras-labs:2.2-r6-SNAPSHOT"
+    } else if (project.jdk == 8) {
+        compile "org.jfxtras:jfxtras-labs:8.0-r1"
+    }
+
+    compile 'org.apache.commons:commons-math3:3.3'
+}
