+++ gradle.properties
@@ -1,0 +1,4 @@
+jacocoVersion = 0.7.4.201502262128
+javadocFooter = 2012-2016 (C) Michael Hoffer <info@michaelhoffer.de>
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
