+++ gradle/idea.gradle
@@ -1,0 +1,33 @@
+// This only gets applied to the root project
+apply plugin: "idea"
+
+idea {
+    project {
+        jdkName "1.8"
+        languageLevel "1.8"
+        ipr {
+            withXml { provider ->
+                def node = provider.asNode()
+                node.component.find { it.'@name' == 'VcsDirectoryMappings' }?.mapping[0].'@vcs' = 'Git'
+
+                def compilerConfiguration = node.component.find { it.'@name' == 'CompilerConfiguration' }
+                compilerConfiguration.remove(compilerConfiguration.annotationProcessing)
+                compilerConfiguration.append(new XmlParser().parseText("""
+    <annotationProcessing>
+      <profile default="true" name="Default" enabled="true">
+        <processorPath useClasspath="true" />
+      </profile>
+    </annotationProcessing>"""))
+
+                def javacSettings = node.component.find { it.@name == 'JavacSettings' }
+                if (!javacSettings) {
+                    node.append(new XmlParser().parseText("""
+          <component name="JavacSettings">
+            <option name="ADDITIONAL_OPTIONS_STRING" value="-Xlint:all -Xlint:deprecation -Xlint:unchecked" />
+          </component>
+          """))
+                }
+            }
+        }
+    }
+}
