+++ gradle/maven-deployment-local.gradle
@@ -1,0 +1,70 @@
+apply from: file('gradle/jdkdetect.gradle')
+
+// sonatype dependencies
+apply plugin: 'maven'
+apply plugin: 'signing'
+
+apply from: file('gradle/version-info.gradle')
+
+
+task javadocJar(type: Jar, dependsOn: javadoc) {
+    classifier = 'javadoc'
+    from 'build/docs/javadoc'
+}
+
+task sourcesJar(type: Jar) {
+    from sourceSets.main.allSource
+    classifier = 'sources'
+}
+
+artifacts {
+    archives jar
+    
+//    archives javadocJar
+//    archives sourcesJar
+}
+
+//signing {
+//    sign configurations.archives
+//}
+
+//
+//uploadArchives {
+//    repositories {
+//        mavenDeployer {
+//            beforeDeployment { MavenDeployment deployment -> signPom(deployment) }
+// 
+//            repository(url: "https://oss.sonatype.org/content/repositories/snapshots/") {
+//              authentication(userName: sonatypeUsername, password: sonatypePassword)
+//            }
+// 
+//            pom.project {
+//               name 'VWorkflows-Core'
+//               packaging 'jar'
+//               description 'VWorkflows is a workflow/graph library with JavaFX bindings'
+//               url 'https://github.com/miho/VWorkflows'
+// 
+//               scm {
+//                   url 'scm:https://github.com/miho/VWorkflows.git'
+//                   connection 'scm:https://github.com/miho/VWorkflows.git'
+//                   developerConnection 'scm:https://github.com/miho/VWorkflows.git'
+//               }
+// 
+//               licenses {
+//                   license {
+//                       name 'The BSD 2-Clause License'
+//                       url 'http://opensource.org/licenses/bsd-license.php'
+//                       distribution 'repo'
+//                   }
+//               }
+// 
+//               developers {
+//                   developer {
+//                       id 'miho'
+//                       name 'Michael Hoffer'
+//                   }
+//               }
+//           }
+//        }
+//    }
+//}
