+++ gradle/version-info.gradle
@@ -1,0 +1,8 @@
+// maven deployment properties
+if (project.jdk == 8) {
+    group = 'eu.mihosoft.vrl.workflow-8.0'
+} else {
+    group = 'eu.mihosoft.vrl.workflow'
+}
+
+version = '0.1-r2-SNAPSHOT'
